benchmarks/model_training/requirements.txt
grpcio==1.45.0
grpcio-tools==1.45.0
grpcio_reflection
scikit-learn
pandas
numpy
joblib